refactoring : entire refactoring, waring catching
