+++ MechanumWheelControl/Decision/Decision_Sub_State.c
@@ -35,7 +35,7 @@
 /***********************************************************************/
 /*Variable*/ 
 /***********************************************************************/
-static uint32  s_sub_state;
+static uint8  s_sub_state;
 sint32 dist;
 /***********************************************************************/
 /*Function*/
