+++ MechanumWheelControl/DeviceDriver/Driver_ToF.c
@@ -10,6 +10,8 @@
 /*Include*/
 /***********************************************************************/
 #include <Driver_ToF.h>
+
+#include <string.h>
 /***********************************************************************/
 /*Define*/
 /***********************************************************************/
