+++ MechanumWheelControl/DeviceDriver/Driver_WheelFL.c
@@ -237,8 +237,8 @@ static inline boolean get_motorFL_direction(void){
 /*------------------------------------------encoder driver Implementations-------------------------------------------*/
 /*********************************************************************************************************************/
 
-IFX_INTERRUPT(encoderFL_chA_isr, 0, ISR_PRIORITY_SCUERU0);
-void encoderFL_chA_isr(void){
+IFX_INTERRUPT(encoderFL_chA_ISR, 0, ISR_PRIORITY_SCUERU0);
+void encoderFL_chA_ISR(void){
     if(IfxPort_getPinState(ENCODER_CHA)) { // rising edge triggered
         if(IfxPort_getPinState(ENCODER_CHB)) {
             s_encoder_data--;
